Added keyboard report stash and restore functions
In order to safely send a mid-cycle report, it should be based on the previous (complete)
report, not the current (incomplete) one. To enable this, I've added functions to "stash"
a copy of the current report and (temporarily) overwrite it with a copy of the previous
one. After a plugin sends it's mid-cycle report, the restore function restores the current
report so that it can be completed normally.

Signed-off-by: Michael Richters <gedankenexperimenter@gmail.com>